feat: Add TotalPages to Paginator struct
