+++ models/Paginator.go
@@ -1,9 +1,10 @@
 package models
 
 type Paginator struct {
-	TotalPages   int         `json:"total_pages"`
-	CurrentPage  int         `json:"current_page"`
-	PreviousPage int         `json:"previous_page"`
-	NextPage     int         `json:"next_page"`
-	Data         interface{} `json:"data"`
+	TotalPages   int         `json:"total_pages,omitempty"`
+	CurrentPage  int         `json:"current_page,omitempty"`
+	PreviousPage int         `json:"previous_page,omitempty"`
+	NextPage     int         `json:"next_page,omitempty"`
+	TotalEntries int         `json:"total_entries,omitempty"`
+	Data         interface{} `json:"data,omitempty"`
 }
